Consolidated user related types, and fixed remaining errors
